try not to clobber original dotfiles

+++ install.sh
@@ -15,8 +15,12 @@ for file in \
   zshenv \
   zshrc
 do
-  echo "Backing up ${HOME}/.${file} to ${backup_dir}/.${file}.bak ..."
-  cp "${HOME}/.${file}" "${backup_dir}/.${file}.bak"
-  echo "Linking ${HOME}/.${file} to ${dotfiles}/${file} ..."
-  ln -sf "${dotfiles}/${file}" "${HOME}/.${file}"
+    if [ -e "${backup_dir}/.${file}" ]; then
+      echo "${backup_dir}/.${file} appears to exist already. Manually symlink your updated file and be careful not to clobber the original file installed on the system."
+    else
+      echo "Backing up ${HOME}/.${file} to ${backup_dir}/.${file}.bak ..."
+      cp "${HOME}/.${file}" "${backup_dir}/.${file}.bak"
+      echo "Linking ${HOME}/.${file} to ${dotfiles}/${file} ..."
+      ln -sf "${dotfiles}/${file}" "${HOME}/.${file}"
+    fi
 done
